Move functions into functions.py
